qa: inline Headers class from laminas-mail
We have a circular dependency scenario between laminas-mail and laminas-mime.
laminas-mail depends on laminas-mime for its MIME parsing capabilities.
laminas-mime decodes headers to a laminas-mail `Headers` instance.
Unfortunately the `Headers` class implements `Iterator`, and, as such, we get deprecations under 8.1 unless the `#[ReturnTypeWillChange]` attribute to affected methods.

The solution here is to temporarily inline the `Headers` class as a test asset until laminas-mail is updated to work on PHP 8.1.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>