Introduce `allow-plugins` config in the composer.json
Changes have been made to meet the requirements of
composer 2.2. Composer 2.2 introduced a new security
layer to prevent executing arbitrary code when the
Composer runs. We have a library
`dealerdirect/phpcodesniffer-composer-installer` which
is executed during the installation of the project.
In this commit, this library was allowed to run during
the installation of the project.

Signed-off-by: danielabyan <danielabyan@gmail.com>