Update phpunit up to `9.5` version
In the old version of PHPUnit there is a bug that
does not allow you to build the project. An error
`Schemas parser error : Failed to locate the main schema
resource at 'vendor/phpunit/phpunit/phpunit.xsd'` is
thrown during the test execution. This bug has been
fixed in version `9.5`. Therefore, the version of the
library has been updated.

Signed-off-by: danielabyan <danielabyan@gmail.com>