Add further assertions when testing multipart alternative messages
This change adds a series of assertions for attachments on
multipart/alternative messages, to help verify that they contain what
they're expected to contain. I felt that it was helpful to have
assertions for attachments (or non text/HTML content) as well, so that
more of the code could be properly tested.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>