Correct a small bug in setting an attachment's filename
Previously, the code only set a filename for a message part if the part
contained a content-disposition header where the disposition type was
set to "attachment" and the filename parameter was set. After re-reading
https://www.w3.org/Protocols/HTTP/Issues/content-disposition.txt, I see
that this was wrong, as the content-disposition header can have the
filename parameter set if the type is attachment or inline.

This commit corrects that earlier mistake.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>